feat(adc): add adc internal temperature sensor example
