+++ bouffalo-hal/src/gpip.rs
@@ -3,5 +3,4 @@
 mod adc;
 mod register;
 
-pub use adc::*;
 pub use register::*;
