+++ examples/peripherals/adc-tsen-demo/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "adc-tsen-demo"
+version = "0.1.0"
+edition = "2024"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bouffalo-hal = { path = "../../../bouffalo-hal", features = ["bl808"] }
+bouffalo-rt = { path = "../../../bouffalo-rt", features = ["bl808-mcu"] }
+panic-halt = "1.0.0"
+embedded-time = "0.12.1"
+riscv = "0.13.0"
+
+[[bin]]
+name = "adc-tsen-demo"
+test = false
